Add quest 5 p2 and p3 solutions

+++ src/main/java/com/pretz/everybodycodes/q5/Main.java
@@ -14,7 +14,7 @@ public class Main {
         }
     }
 
-    private static int solve(String[] args) throws IOException {
+    private static long solve(String[] args) throws IOException {
         return Objects.requireNonNull(
                         SolverBuilder.build(Integer.parseInt(args[0]))
                 )
